[FEATURE] Include preview link in notification
Until now there haven't been a preview link in the notification mails.

With this feature there are two new markers that can be used in the
notification template. One marker for the workspace preview link and
one marker for the splitted workspace preview link.
Additionally this feature adds a hook, so it is possible for a
developer to add severall additional markers.

The preview link for the splitted workspace view goes across the
backend login screen. If an user has no valid backend session he
can simply login and after an successfull login he will be redirected
to the desired worskpace view. If the user already logged in to the
backend he becomes directly the workspace module without any new
login.

If the recipient of the review email was last time logged in into the
live WS. The workspace will now forced (if the permission allows) to
the desired WS where the changes are present. Otherwise it's not possible
to get the splitted preview screen running.

Change-Id: Ifec84d7475aa111cc55a0fb2915d3ddfa6c5fcc4
Resolves: #13518