[BUGFIX] Avoid SQL error when module is opened on moved page
In case the backend module is opened on a page which was moved
within the current workspace there's a SQL error caused because
the processing for the relevant "Pagetree" runs empty.
Adding the current page to the processing can never break this part
because it will always be part of the result anyways.

Change-Id: I57d7740a4c436bf3fd44b8cc93d4ae1d9e9529b9
Resolves: #13635