[TASK] Use alternative selector for many workspaces
The toolbar of the workspace module visualizes each workspace
in a separate tab. On having many workspaces, the toolbar is
extended to multiple lines which actually reduces the available
viewport of the grid panel.

This change introduces a modified TabPanel that shrinks if the
available width is undershot. In this case an additional menu
is rendered to the right side of the panel to visualize the
remaining items in a vertical list.

Resolves: #49689
Releases: 6.2
Change-Id: I85c8b8134dfe27a4dc32c5d70a158cca41763c41
Reviewed-on: https://review.typo3.org/21920
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader