[FEATURE] Enable all workspaces tab in workspace module
The "all workspaces" tab currently is only available for admin
users. However, there's no need to disallow regular editors to
use this feature as well - access will be granted to workspaces
that have been defined for each particular user or group.

Besides that, actions from the "all workspaces" view need to be
handled in a special way since there is no correct
BE_USER->workspace value assigned in that case, but working with
versioned records requires to have a correct scope defined.

Change-Id: I89bfee3ad0b2ce1a5f3c12a1bd91342472b12d3d
Resolves: #50223
Releases: 6.2
Reviewed-on: https://review.typo3.org/22445
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader