[BUGFIX] Workspace preview links for multiple languages
The workspace preview link will only be generated for one language at
a time, so it might require some effort to generate preview links for
multiple languages.

For convenience of editors, preview links will now be generated for
all available translations of a certain page in a batch.

Resolves: #20852
Resolves: #12970
Releases: master, 6.2
Change-Id: I6de6f2c7b741e238fc373f0756bc29f0100a19f0
Reviewed-on: http://review.typo3.org/26993
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Daniel Maier <dani-maier@gmx.de>
Tested-by: Daniel Maier <dani-maier@gmx.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>