[TASK] Add functional test cases to DataHandler (workspace)
In addition to the functional test cases for data handling in
live, these test cases re-use the test scenarios to check the
behaviour in a particular workspace.

The current assertions only check the current behaviour. Parts
that are identified as faults are marked out with the ignore
operator "\*" in the CSV files.

Thus, these tests shall be used to compare the current state
with the target state in combination with accordant bug-fixes
for workspace data handling.

Resolves: #55563
Releases: 6.2
Change-Id: I5e35f0f450093416e5165fc399da3b454715a5d7
Reviewed-on: https://review.typo3.org/27241
Reviewed-by: Peter Kuehn
Tested-by: Peter Kuehn
Reviewed-by: Jigal van Hemert
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader