[BUGFIX] Discard vs. Delete workspace changes in tests
Discarding workspace changes works with the version command
"clearWSID" which might come down to a "delete" action if the
versioned records have a delete or new placeholder (which is
the case for copied or localized elements).
However, the functional tests need to use the "clearWSID"
instead of the "delete" command.

Resolves: #56111
Releases: 6.2
Change-Id: Id10413048a68abdb5ef05bb6e20e6cb6460f9742
Reviewed-on: https://review.typo3.org/27712
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader