[TASK] Visualize nested data structures in workspace module
The current workspace module only has two levels (page and the
accordant records). Nested record sets like tt_content ->
sys_file_reference (any "text with image" content element) are
not recognized as dependent and need to be published separately.

The GridDataService is extended to determine the the accordant
nested record sets and provides additional data for parent and
child scenarios. The ExtJS view components are extended to take
care of nested record sets and to handle expand and collapse
events on these kind of record collections.

Resolves: #55349
Releases: 6.2
Change-Id: I93ca187c3997bf7f4cdadefd741be2541aef5ae4
Reviewed-on: https://review.typo3.org/28701
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader