[TASK] Re-work/simplify copyright header in PHP files - Part 9
This patch replaces the copyright/license header in PHP files with a
new, simplified one. The new header does not show the year figure, nor
an author name, and refers to the LICENSE.txt file for the full
copyright information. License is: GPL2 or any later version.

This is a multi-part commit due to the huge number of changed files.
See issue #59785 for further details.

Resolves: #59785
Releases: 6.3, 6.2
Change-Id: If6a4795e7f2841474241d236d8ed33baceee0d2e
Reviewed-on: https://review.typo3.org/31030
Reviewed-by: Markus Klein
Tested-by: Markus Klein