[CLEANUP] Add missing empty lines before classes' closing curly braces
Most of TYPO3's PHP files have an empty line before a class' closing curly
brace. This commit adds this empty line (if it is missing) before a class'
closing curly brace.

This is not part of the TYPO3 CMS CGL but increases consistency in this
area.

This regular expression was used to add the empty lines:

Search:
(}\n)(})

Replace:
$1\n$2

Resolves: #63942
Release: master
Change-Id: Icd6547322ac8b7310956d347fd73c90bde2bbfa7
Reviewed-on: http://review.typo3.org/35527
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>