[!!!][TASK] Remove option $TYPO3_CONF_VARS[BE][accessListRenderMode]
With the recent overhaul to optimize the selection
of access lists (permissions) in BE user / BE groups
via checkbox,
the option $TYPO3_CONF_VARS[BE][accessListRenderMode]
which was set to "checkbox" already, can be removed
as it only serves as a shorthand function
to override TCA settings for be_users and be_groups.

In the front of having strong defaults and a simpler
option list, the option is removed and all values
within be_users/be_groups are set to
renderMode=checkbox directly.

Resolves: #64226
Releases: master
Change-Id: Ic6076fb44b6bb6874041a22ee30069457a17d5a0
Reviewed-on: http://review.typo3.org/35939
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>