[BUGFIX] Non-admins cannot use mass actions in a workspace
Non-admins cannot use the mass actions "publish", "swap" or
"discard" in a workspace. The reason is, that for non-admins
the mount-page tree is traversed for a given recursion depth
which is not the case for privileged admin users. The default
value for the depth value is zero (0), that's why no pages,
except the mount-points, are processed.

Resolves: #59226
Releases: master, 6.2
Change-Id: I8d1e2032f72fa977cfcaf44978300d36bef9be23
Reviewed-on: https://review.typo3.org/44417
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>