[TASK] Move workspaces preview hook to EXT:workspaces
Move the preview functionality (of sys_preview) from EXT:version to
EXT:workspaces as it resolves solely functionality to preview
a page within a workspace with the link.

A class map alias ensures that all functionality is still given
and non-breaking, however the SQL information was moved to
EXT:workspaces, so it breaks if only EXT:version is installed and
e.g. sys_preview was used for something unrelated to
the preview functionality.

Resolves: #78434
Releases: master
Change-Id: If5ca8cf7c3a983b84274778913860764d4c3a803
Reviewed-on: https://review.typo3.org/50330
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>