[BUGFIX] Make checks in localize compatible with new Localization Wizard
The checks for language and localization-parent in DataHandler's localize
method are made compatible with the options which the new localization wizard
allows; more specifically allowing translations from other languages than the
default language (but still pointing to a default record).

This makes it also possible to make correct use of the localize command
instead of the copy command now when translating from other languages than
the default language.

Resolves: #78059
Releases: master, 7.6
Change-Id: I01647d759a745abbbf0fa34a16b20f3085ef91ea
Reviewed-on: https://review.typo3.org/50022
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>