[!!!][TASK] Workspaces: Always consider references
Previously, it was possible to disable the handling
of references / child-parent relations in workspaces
on publishing, due to legacy functionality this was
an option introduced back in v4.

The behaviour (option switch via TSconfig) was
enabled by default on all installations.

It automatically resolves dependencies
(= publishing dependencies in relations automatically),
having this option disabled could lead to side effects
in the past.

Resolves: #82701
Releases: master
Change-Id: I4f152fd337e4c405eaa0f1386601629a688a05d4
Reviewed-on: https://review.typo3.org/54319
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Henning Liebe <h.liebe@neusta.de>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>