[BUGFIX] Better sorting localizing records from non-default language
Using the "translate" button in page module and localizing
records from non-default language may lead to funny content
element sorting in the target language.

This has been mitigated with probably #85300 and #84951
a while ago already, but there still exists a codewise
misbehavior in DataHandler::getPreviousLocalizedRecordUid().

The patch streamlines the method and applies a fix so
records calculate the correct "before" uid even in
"localization of a localization" context.

Resolves: #83079
Resolves: #83469
Related: #85300
Related: #84951
Releases: master, 11.5
Change-Id: I03c566054bd4939fed96206f62633b2abe4f195f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72254
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>