[TASK] Move workspace preview user to custom class
Instead of faking an existing backend user, the workspace preview
functionality (= from a link) should use an anonymous read-only
use which has only access to the workspace and the page, in order
to remove the hacks to allow read-access for existing backend users.

This way, the hook code is getting cleaner and easier to read, and also
easier to debug.

Resolves: #84039
Releases: master
Change-Id: Ia69d66ce25af48b86104ff724f2a3e877aa3a813
Reviewed-on: https://review.typo3.org/55888
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>