[FEATURE] Move workspaces scheduler tasks to native symfony commands
The previously available scheduler tasks for auto-publishing and cleaning up
preview links have been moved to a symfony/command, allowing them to be
scheduled, but also triggered manually on the command line without having
to create a task record in scheduler.

The old tasks have been deprecated, as there is the possibility to create
tasks out of symfony commands as well.

The AutoPublishService, which only acted as a wrapper, but had a major
design flaw has been deprecated as well.

The scheduler tasks will be removed in TYPO3 v10. For TYPO3 v9, their
labels have been updated, and they will trigger a deprecation notice.

Resolves: #86001
Releases: master
Change-Id: I9c1f1e31b6ff59d02b20ffc2df805e645bb3f675
Reviewed-on: https://review.typo3.org/58056
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>