[BUGFIX] Prevent PHP type error and stale language cache
PHP is raising an error if $pageId cannot be resolved to an
integer value as required by method signature.

Additionally, an internally cached array is built based on a
dynamic value but the internal cache is not keyed by this
value or flushed when the value changes.

This patch:

* Casts an unresolved page ID to integer zero.
* Removes the internal cache of languages since the array
  now depends entirely on a unique input value, $pageId.

Resolves: #89692
Resolves: #90073
Releases: master, 9.5
Change-Id: I3d74b4c7575375aae23a7e2473100104c4700966
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62974
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Alexander Schnitzler <git@alexanderschnitzler.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Alexander Schnitzler <git@alexanderschnitzler.de>
Reviewed-by: Benni Mack <benni@typo3.org>