[FEATURE] Add a stage filter to the workspace module
The workspace module gets an additional stage filter
to reduce the amount of changes a user is facing. This
is very helpful if you are dealing with a lot of changes
and want to see only records which are in a certain stage
e.g. "Ready to publish".

Resolves: #91021
Releases: master
Change-Id: I80a3fa5c00ff302c9b2257370bdeda328033c204
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64197
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>