[BUGFIX] Display human-readable preview of FlexForm values
This change transforms the processed record of FlexForms
into a human-readable format.

This is especially useful when comparing changes in the
history module or the workspaces diff view.

The diff needs to be formatted with line breaks and tabs as
the output is passed through "htmlspecialchars" before it
will be displayed.

The white-space css setting needs to be changed to
"pre-wrap", because this is the only setting where line
breaks and tabs are kept and the text will still be
word-wrapped.

The old view is used as a fallback, if no record is
available (import preview).

Resolves: #91148
Releases: main, 11.5
Change-Id: Iab1dfb59f98dc712b8984a8087525634b5ec4ebd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64260
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>