[BUGFIX] Deleting a localized page deletes its localized records
This fixes an issue that has been in the core "since ever":
Have a localized page with localized records on it. If the
page localization is deleted (for instance via list module),
its localized records are NOT marked as deleted. They are
not shown anymore since the localized page record is gone,
but they're still there. If later a localization of this page
with the same sys_language_uid is created again, the records
magically reappear.
The patch changes DataHandler->deleteSpecificPage() to delete
localized records of localized pages correctly for the
sys_language_uid in question if the table is localization
aware.
An edge case within workspace is handled to suppress the
cascading delete while swapping pages.

Resolves: #90447
Releases: master, 10.4, 9.5
Change-Id: Iea44deeea929f165c717d17e6b997ff6bb829847
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65067
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>