[BUGFIX] No bogus log entries when discarding workspace records
Selecting a page and records on the page to be discarded in the
workspace module can lead to log entries that records could not
be deleted.
This can happen when discarding (deleting) a page that only exists
in the workspace already has deleted records on the page. The explicit
call to delete single records then creates this log.
The solution is to ask the DataHandler if the record in question
has already been deleted to avoid deletion in this case.
Additionally, the log entry now contains the table:uid to
make it more useful in case it pops up again.

Resolves: #61719
Releases: master, 10.4
Change-Id: I8fee34fad2484e9e656bcb5ff611a91374127d94
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65106
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>