[BUGFIX] Add more log info when discarding workspace records fails
Discarding records in the workspace module may lead to
'You do not have edit access' log entries. The issue is hard
to reproduce without the table name and uid in the log.
The patch changes the log entry to contain this info.
This is similar to the log entry change for issue #61719.

Resolves: #69310
Related: #61719
Related: #91881
Releases: master, 10.4
Change-Id: I0da0d13cb6be193f17835bedc1027dbf6559a796
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65105
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>