[BUGFIX] Skip processing of inline children which aren't workspace-aware
When dealing with inline foreign field setups in workspaces
and having a child table not workspace aware, the DataHandler
still tries to create workspace copies of the not workspace
aware child.

In v9, it sets pid=-1 for those records, which succeeds if the
pid db field is signed. In this case a dead record of the
child is created. If the field is unsigned, the insertion fails
and an error message is created.

In v10 and master pid=-1 is gone and the real pid is set, the
record is created and becomes a 1:1 duplicate in live workspace.

The patch adds an according check to the DataHandler in
workspace context to skip not workspace aware children.

Resolves: #91680
Releases: master, 10.4, 9.5
Change-Id: Ifb8183f25e69a75ecc5b5ebac3c9a57c03265e7f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65230
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: David Steeb <david.steeb@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: David Steeb <david.steeb@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>