[!!!][TASK] Remove support for swapping instead of publishing workspace records
The functionality of "exchanging" a versioned record in a workspace with
a live version, is not used, and has certain downsides (e.g. it is not clear
what exchange happened, not event stream to understand when
was what changed, and especially when swapping single/selected items
multiple times it gets really messy). Also, swapping is conceptually
not working (when activated) for newly created records and leaves an
inconsistent state. For these reasons, the feature is dropped.

What was removed:
* The option sys_workspace.swap_modes
* The option to only swap on auto-publish
* The option to swap in workspace module

In the future it would be really beneficial to rename internal "swap"
variables to "publish" to streamline this logic.

Resolves: #92206
Releases: master
Change-Id: Ie153a890e5d0dc6ac0b8707a2739c89922e017d7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65318
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>