[TASK] Default workspaces_perms=0 for new be_users
Field workspace_perms=1 for backend users or groups enables
editing of the live workspace.
When ext:workspaces is loaded, we can assume the extension
is actively used in the instance. If so, many backend users
most likely should NOT have live edit access and the flag
is probably given to only a small group of editors by assigning
a 'chief editor' group or similar that carries the flag.
The patch sets workspace_perms for new backend users by
default to 0 when ext:workspaces is loaded, making it more
unlikely admins accidentially give live edit access to users.

Resolves: #92134
Releases: master, 10.4
Change-Id: Id69a60c732e3dc9a55d8dd861ca1a89792f49af1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65506
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Benni Mack <benni@typo3.org>