[BUGFIX] Ensure workspace's new placeholder keeps sorting+pid on moving
When a newly created record in a workspace is added, two records
are created:

  * t3ver_state=-1 (the versioned record NEW_PLACEHOLDER_VERSION which contains the data)
  * t3ver_state=1 (the placeholder NEW_PLACEHOLDER)

When the record is now moved (= sorting or pid is changed) then
the versioned record was previously not updated sorting or pid values.

This change keeps both records in sync making it easier to have a
consistent setup. This is helpful when working on versioned records
(and not on the pairs = overlays) directly.

Resolves: #92182
Releases: master, 10.4
Change-Id: Idc7632b12cc03a04eac1430f84cf67bf1ec24e73
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65561
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>