[TASK] sys_refindex testing in DataHandler functional tests
To get a feeling of current state of sys_refindex handling in the
core, the DataHandler functional tests now assert the state of
this table.
This is a pre-patch to fix reference index details later, so
workspaces and FAL can eventually become friends.

typo3/testing-framework needs a raise to deal with sys_refindex:
composer require --dev typo3/testing-framework:^6.4.2

A couple of tests are left out since they trigger a bug
if using postresql. Those are listed in #92190 and need a
separate patch.

Change-Id: I6f83971687409e31506cea0177ddf64d707307ad
Resolves: #92189
Related: #92190
Releases: master, 10.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65544
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>